+++ hooks/src/lib/validators/validatePassword.ts
@@ -1,0 +1,14 @@
+export const validatePassword = (
+  value: string,
+  passwordLength: number,
+): string => {
+  if (value !== '' && Number.isNaN(Number(value))) {
+    return '숫자만 입력 가능합니다.';
+  }
+
+  if (value !== '' && value.length !== passwordLength) {
+    return `${passwordLength}자리를 입력해주세요.`;
+  }
+
+  return '';
+};
